+++ bridge/bridge.go
@@ -2,7 +2,7 @@ package bridge
 
 import (
 	"errors"
-	"log"
+	log "github.com/pipedrive/registrator/logger"
 	"net"
 	"net/url"
 	"os"
@@ -11,8 +11,6 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-
-	"syscall"
 
 	dockerapi "github.com/fsouza/go-dockerclient"
 )
@@ -59,16 +57,6 @@ func (b *Bridge) Add(containerId string) {
 }
 
 func (b *Bridge) SetupSigtermBehavior(behavior string, msg *dockerapi.APIEvents, ttl int, checkStatus string) {
-	signal := -1
-	i, err := strconv.Atoi(msg.Actor.Attributes["signal"])
-	if err == nil {
-		signal = i
-	}
-
-	if syscall.Signal(signal) != syscall.SIGTERM {
-		return
-	}
-
 	switch behavior {
 	case "deregister":
 		b.remove(msg.ID, true)
@@ -203,6 +191,9 @@ func (b *Bridge) Sync(quiet bool) {
 
 func (b *Bridge) add(containerId string, quiet bool) {
 	if d := b.deadContainers[containerId]; d != nil {
+		if log.IsVerbose() {
+			log.Debugf("Found dead container %s", containerId[:12])
+		}
 		b.services[containerId] = d.Services
 		delete(b.deadContainers, containerId)
 	}
@@ -215,6 +206,9 @@ func (b *Bridge) add(containerId string, quiet bool) {
 
 	if filtersLength := len(b.config.ContainersFilters); filtersLength > 0 {
 		filters := b.config.ContainersFilters.WithContainerId(containerId)
+		if log.IsVerbose() {
+			log.Debugf("Applying filters for container %s - %s", containerId[:12], filters)
+		}
 		containersList, err := b.docker.ListContainers(dockerapi.ListContainersOptions{Filters: filters})
 		if err == nil && len(containersList) == 0 {
 			log.Printf("Ignoring container: %s as it doesn't match \"%s\" filters ", containerId[:12], b.config.ContainersFilters.String())
@@ -233,13 +227,33 @@ func (b *Bridge) add(containerId string, quiet bool) {
 
 	// Extract configured host port mappings, relevant when using --net=host
 	for port, _ := range container.Config.ExposedPorts {
-		published := []dockerapi.PortBinding{ {"0.0.0.0", port.Port()}, }
+		published := []dockerapi.PortBinding{{"0.0.0.0", port.Port()},}
+		if log.IsVerbose() {
+			log.Debugf("Found ExposedPort for %s - \"%s\"", container.ID[:12], port.Port())
+			log.Separator()
+		}
 		ports[string(port)] = servicePort(container, port, published)
+		if log.IsVerbose() {
+			log.Separator()
+			log.Debugf("Adding ExposedPort for %s - \"%+v\"", container.ID[:12], ports[string(port)])
+			log.Separator()
+			log.Separator()
+		}
 	}
 
 	// Extract runtime port mappings, relevant when using --net=bridge
 	for port, published := range container.NetworkSettings.Ports {
+		if log.IsVerbose() {
+			log.Debugf("Found Bridged port for %s - \"%s\" with PortBindings \"%+v\"", container.ID[:12], port.Port(), published)
+			log.Separator()
+		}
 		ports[string(port)] = servicePort(container, port, published)
+		if log.IsVerbose() {
+			log.Separator()
+			log.Debugf("Adding ports for %s from bridge network %+v", container.ID[:12], ports[string(port)])
+			log.Separator()
+			log.Separator()
+		}
 	}
 
 	if len(ports) == 0 && !quiet {
@@ -267,6 +281,9 @@ func (b *Bridge) add(containerId string, quiet bool) {
 			}
 			continue
 		}
+		if log.IsVerbose() {
+			log.Debugf("Registering service for %s: %+v", container.ID[:12], service)
+		}
 		err := b.registry.Register(service)
 		if err != nil {
 			log.Println("register failed:", service, err)
@@ -278,6 +295,9 @@ func (b *Bridge) add(containerId string, quiet bool) {
 }
 
 func (b *Bridge) newService(port ServicePort, isgroup bool) *Service {
+	if log.IsVerbose() {
+		log.Debugf("Preparing service information from %s for port %s", port.ContainerID[:12], port.ExposedPort)
+	}
 	container := port.container
 	defaultName := strings.Split(path.Base(container.Config.Image), ":")[0]
 
@@ -293,11 +313,24 @@ func (b *Bridge) newService(port ServicePort, isgroup bool) *Service {
 		}
 	}
 
-	if b.config.HostIp != "" {
-		port.HostIP = b.config.HostIp
+	if log.IsVerbose() {
+		log.Debugf("Service %s host IP is %s:", port.ContainerID[:12], port.HostIP)
 	}
 
+	if b.config.HostIp != "" {
+		port.HostIP = b.config.HostIp
+		if log.IsVerbose() {
+			log.Debugf("Service %s forced IP to %s:", port.ContainerID[:12], b.config.HostIp)
+		}
+	}
+
+	if log.IsVerbose() {
+		log.Debugf("Extracting meta data for %s for port %s", port.ContainerID[:12], port.ExposedPort)
+	}
 	metadata, metadataFromPort := serviceMetaData(container.Config, port.ExposedPort)
+	if log.IsVerbose() {
+		log.Debugf("Extracted meta data for %s is:(general): %+v; port-specific: %+v", port.ContainerID[:12], metadata, metadataFromPort)
+	}
 
 	ignore := mapDefault(metadata, "ignore", "")
 	if ignore != "" {
@@ -322,6 +355,9 @@ func (b *Bridge) newService(port ServicePort, isgroup bool) *Service {
 		p, _ = strconv.Atoi(port.HostPort)
 	}
 	service.Port = p
+	if log.IsVerbose() {
+		log.Debugf("Service for %s exposed IP address us %s", port.ContainerID[:12], service.IP)
+	}
 
 	if b.config.UseIpFromLabel != "" {
 		containerIp := container.Config.Labels[b.config.UseIpFromLabel]
@@ -356,13 +392,15 @@ func (b *Bridge) newService(port ServicePort, isgroup bool) *Service {
 		}
 	}
 
+	service.Tags = combineTags(
+		mapDefault(metadata, "tags", ""),
+		b.config.ForceTags,
+		hostname+"-"+container.Config.Hostname,
+		EvaluateTemplateTags(&b.config.TemplateTags, container))
+
 	if port.PortType == "udp" {
-		service.Tags = combineTags(
-			mapDefault(metadata, "tags", ""), b.config.ForceTags, "udp", hostname+"-"+container.Config.Hostname)
+		service.Tags = append(service.Tags, "udp")
 		service.ID = service.ID + ":udp"
-	} else {
-		service.Tags = combineTags(
-			mapDefault(metadata, "tags", ""), b.config.ForceTags, hostname+"-"+container.Config.Hostname)
 	}
 
 	id := mapDefault(metadata, "id", "")
@@ -383,6 +421,9 @@ func (b *Bridge) newService(port ServicePort, isgroup bool) *Service {
 func (b *Bridge) setupTtlHealthCheck(containerId string, ttlHealthCheck *TtlHealthCheck) {
 	b.Lock()
 	defer b.Unlock()
+	if log.IsVerbose() {
+		log.Debugf("Setting up TTL health check for %s", containerId[:12])
+	}
 	setupAllHealthChecks := func(services []*Service) {
 		for _, service := range services {
 			err := b.registry.SetupHealthCheck(service, ttlHealthCheck)
@@ -394,11 +435,18 @@ func (b *Bridge) setupTtlHealthCheck(containerId string, ttlHealthCheck *TtlHeal
 		}
 	}
 	setupAllHealthChecks(b.services[containerId])
+	if d := b.deadContainers[containerId]; d != nil {
+		setupAllHealthChecks(d.Services)
+	}
 }
 
 func (b *Bridge) remove(containerId string, deregister bool) {
 	b.Lock()
 	defer b.Unlock()
+
+	if log.IsVerbose() {
+		log.Debugf("Going to remove %s services", containerId[:12])
+	}
 
 	if deregister {
 		deregisterAll := func(services []*Service) {